MAINT - Moving negative path tests to separate files
