+++ ctk/idp/src/main/kotlin/org/codice/compliance/web/sso/error/PostSSOErrorTest.kt
@@ -11,10 +11,9 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.compliance.web.sso
+package org.codice.compliance.web.sso.error
 
 import com.jayway.restassured.RestAssured
-import com.jayway.restassured.RestAssured.given
 import com.jayway.restassured.response.Response
 import de.jupf.staticlog.Log
 import io.kotlintest.specs.StringSpec
@@ -26,25 +25,21 @@ import org.codice.compliance.SAMLProfiles_4_1_4_1_a
 import org.codice.compliance.SAMLProfiles_4_1_4_1_b
 import org.codice.compliance.debugPrettyPrintXml
 import org.codice.compliance.debugWithSupplier
-import org.codice.compliance.saml.plugin.IdpSSOResponder
-import org.codice.compliance.utils.TestCommon
 import org.codice.compliance.utils.TestCommon.Companion.AUTHN_REQUEST
 import org.codice.compliance.utils.TestCommon.Companion.EXAMPLE_RELAY_STATE
+import org.codice.compliance.utils.TestCommon.Companion.ID
 import org.codice.compliance.utils.TestCommon.Companion.INCORRECT_ACS_URL
 import org.codice.compliance.utils.TestCommon.Companion.INCORRECT_DESTINATION
+import org.codice.compliance.utils.TestCommon.Companion.RELAY_STATE_GREATER_THAN_80_BYTES
 import org.codice.compliance.utils.TestCommon.Companion.REQUESTER
+import org.codice.compliance.utils.TestCommon.Companion.SP_ISSUER
 import org.codice.compliance.utils.TestCommon.Companion.acsUrl
 import org.codice.compliance.utils.TestCommon.Companion.authnRequestToString
-import org.codice.compliance.utils.TestCommon.Companion.getServiceProvider
-import org.codice.compliance.utils.decorate
+import org.codice.compliance.utils.TestCommon.Companion.parseErrorResponse
 import org.codice.compliance.verification.binding.BindingVerifier
 import org.codice.compliance.verification.core.CoreVerifier
-import org.codice.compliance.verification.core.responses.AuthnRequestProtocolResponseVerifier
 import org.codice.compliance.verification.profile.ProfilesVerifier
-import org.codice.compliance.verification.profile.SingleSignOnProfileVerifier
 import org.codice.security.saml.SamlProtocol
-import org.codice.security.saml.SamlProtocol.Binding.HTTP_POST
-import org.codice.security.saml.SamlProtocol.POST_BINDING
 import org.codice.security.sign.Encoder
 import org.codice.security.sign.SimpleSign
 import org.joda.time.DateTime
@@ -54,7 +49,7 @@ import org.opensaml.saml.saml2.core.impl.IssuerBuilder
 import org.opensaml.saml.saml2.core.impl.NameIDPolicyBuilder
 import org.opensaml.saml.saml2.core.impl.SubjectBuilder
 
-class PostLoginTest : StringSpec() {
+class PostSSOErrorTest : StringSpec() {
     companion object {
 
         /** Sets up positive path tests.
@@ -63,18 +58,18 @@ class PostLoginTest : StringSpec() {
         private fun createValidAuthnRequest(): String {
             val authnRequest = AuthnRequestBuilder().buildObject().apply {
                 issuer = IssuerBuilder().buildObject().apply {
-                    value = TestCommon.SP_ISSUER
+                    value = SP_ISSUER
                 }
-                assertionConsumerServiceURL = acsUrl[HTTP_POST]
-                id = TestCommon.ID
+                assertionConsumerServiceURL = acsUrl[SamlProtocol.Binding.HTTP_POST]
+                id = ID
                 version = SAMLVersion.VERSION_20
                 issueInstant = DateTime()
-                destination = Common.getSingleSignOnLocation(POST_BINDING)
-                protocolBinding = POST_BINDING
+                destination = Common.getSingleSignOnLocation(SamlProtocol.POST_BINDING)
+                protocolBinding = SamlProtocol.POST_BINDING
                 nameIDPolicy = NameIDPolicyBuilder().buildObject().apply {
                     allowCreate = true
                     format = SAML2Constants.NAMEID_FORMAT_PERSISTENT
-                    spNameQualifier = TestCommon.SP_ISSUER
+                    spNameQualifier = SP_ISSUER
                 }
                 SimpleSign().signSamlObject(this)
             }
@@ -85,99 +80,19 @@ class PostLoginTest : StringSpec() {
         }
 
         private fun sendAuthnRequest(encodedRequest: String): Response {
-            return given()
+            return RestAssured.given()
                     .urlEncodingEnabled(false)
                     .body(encodedRequest)
                     .contentType("application/x-www-form-urlencoded")
                     .log()
                     .ifValidationFails()
                     .`when`()
-                    .post(Common.getSingleSignOnLocation(POST_BINDING))
+                    .post(Common.getSingleSignOnLocation(SamlProtocol.POST_BINDING))
         }
     }
 
     init {
         RestAssured.useRelaxedHTTPSValidation()
-
-        "POST AuthnRequest Test" {
-            Log.debugWithSupplier { "POST AuthnRequest Test" }
-            val encodedRequest = Encoder.encodePostMessage(createValidAuthnRequest())
-            val response = sendAuthnRequest(encodedRequest)
-            BindingVerifier.verifyHttpStatusCode(response.statusCode)
-
-            val idpResponse = getServiceProvider(IdpSSOResponder::class)
-                    .getPostResponse(response).decorate()
-            // TODO When DDF is fixed to return a POST SSO response, uncomment this line
-            // SingleSignOnProfileVerifier.verifyBinding(idpResponse)
-            idpResponse.bindingVerifier().verify()
-
-            val responseDom = idpResponse.responseDom
-            AuthnRequestProtocolResponseVerifier(responseDom, TestCommon.ID, acsUrl[HTTP_POST])
-                    .verify()
-            SingleSignOnProfileVerifier(responseDom, acsUrl[HTTP_POST]).verify()
-        }
-
-        "POST AuthnRequest With Relay State Test" {
-            Log.debugWithSupplier { "POST AuthnRequest With Relay State Test" }
-            val encodedRequest = Encoder.encodePostMessage(
-                    createValidAuthnRequest(), EXAMPLE_RELAY_STATE)
-            val response = sendAuthnRequest(encodedRequest)
-            BindingVerifier.verifyHttpStatusCode(response.statusCode)
-
-            val idpResponse = getServiceProvider(IdpSSOResponder::class)
-                    .getPostResponse(response).decorate().apply {
-                        isRelayStateGiven = true
-                    }
-            // TODO When DDF is fixed to return a POST SSO response, uncomment this line
-            // SingleSignOnProfileVerifier.verifyBinding(idpResponse)
-            idpResponse.bindingVerifier().verify()
-
-            val responseDom = idpResponse.responseDom
-            AuthnRequestProtocolResponseVerifier(responseDom, TestCommon.ID, acsUrl[HTTP_POST])
-                    .verify()
-            SingleSignOnProfileVerifier(responseDom, acsUrl[HTTP_POST]).verify()
-        }
-
-        "POST AuthnRequest Without ACS Url Test" {
-            Log.debugWithSupplier { "POST AuthnRequest Without ACS Url Test" }
-            val authnRequest = AuthnRequestBuilder().buildObject().apply {
-                issuer = IssuerBuilder().buildObject().apply {
-                    value = TestCommon.SP_ISSUER
-                }
-                id = TestCommon.ID
-                version = SAMLVersion.VERSION_20
-                issueInstant = DateTime()
-                destination = Common.getSingleSignOnLocation(POST_BINDING)
-                protocolBinding = POST_BINDING
-                nameIDPolicy = NameIDPolicyBuilder().buildObject().apply {
-                    allowCreate = true
-                    format = SAML2Constants.NAMEID_FORMAT_PERSISTENT
-                    spNameQualifier = TestCommon.SP_ISSUER
-                }
-                SimpleSign().signSamlObject(this)
-            }
-
-            val authnRequestString = authnRequestToString(authnRequest)
-            authnRequestString.debugPrettyPrintXml(AUTHN_REQUEST)
-
-            val encodedRequest = Encoder.encodePostMessage(
-                    authnRequestString,
-                    EXAMPLE_RELAY_STATE)
-
-            val response = sendAuthnRequest(encodedRequest)
-            BindingVerifier.verifyHttpStatusCode(response.statusCode)
-
-            val idpResponse = getServiceProvider(IdpSSOResponder::class)
-                    .getPostResponse(response).decorate()
-            // TODO When DDF is fixed to return a POST SSO response, uncomment this line
-            // SingleSignOnProfileVerifier.verifyBinding(idpResponse)
-            idpResponse.bindingVerifier().verify()
-
-            val responseDom = idpResponse.responseDom
-            AuthnRequestProtocolResponseVerifier(responseDom, TestCommon.ID, acsUrl[HTTP_POST])
-                    .verify()
-            SingleSignOnProfileVerifier(responseDom, acsUrl[HTTP_POST]).verify()
-        }
 
         // Negative Path Tests
         "POST AuthnRequest With Relay State Greater Than 80 Bytes Test" {
@@ -185,16 +100,16 @@ class PostLoginTest : StringSpec() {
                 "POST AuthnRequest With Relay State Greater Than 80 Bytes Test"
             }
             val encodedRequest = Encoder.encodePostMessage(
-                    createValidAuthnRequest(), TestCommon.RELAY_STATE_GREATER_THAN_80_BYTES)
+                    createValidAuthnRequest(), RELAY_STATE_GREATER_THAN_80_BYTES)
             val response = sendAuthnRequest(encodedRequest)
 
-            val idpResponse = TestCommon.parseErrorResponse(response)
+            val idpResponse = parseErrorResponse(response)
             idpResponse.bindingVerifier().verifyError()
 
             val responseDom = idpResponse.responseDom
             CoreVerifier(responseDom).verifyErrorStatusCode(
                     samlErrorCode = SAMLBindings_3_5_3_a,
-                    expectedStatusCode = TestCommon.REQUESTER)
+                    expectedStatusCode = REQUESTER)
         }.config(enabled = false)
 
         "Empty POST AuthnRequest Test" {
@@ -205,15 +120,17 @@ class PostLoginTest : StringSpec() {
             val authnRequestString = authnRequestToString(authnRequest)
             authnRequestString.debugPrettyPrintXml(AUTHN_REQUEST)
 
-            val encodedRequest = Encoder.encodePostMessage(authnRequestString, EXAMPLE_RELAY_STATE)
+            val encodedRequest = Encoder.encodePostMessage(authnRequestString,
+                    EXAMPLE_RELAY_STATE)
             val response = sendAuthnRequest(encodedRequest)
             BindingVerifier.verifyHttpStatusCode(response.statusCode)
 
-            val idpResponse = TestCommon.parseErrorResponse(response)
+            val idpResponse = parseErrorResponse(response)
             idpResponse.bindingVerifier().verifyError()
 
             val responseDom = idpResponse.responseDom
-            CoreVerifier(responseDom).verifyErrorStatusCode(SAMLProfiles_4_1_4_1_a, REQUESTER)
+            CoreVerifier(responseDom).verifyErrorStatusCode(SAMLProfiles_4_1_4_1_a,
+                    REQUESTER)
             ProfilesVerifier(responseDom).verifyErrorResponseAssertion()
         }.config(enabled = false)
 
@@ -221,13 +138,13 @@ class PostLoginTest : StringSpec() {
             Log.debugWithSupplier { "POST AuthnRequest With Empty Subject Test" }
             val authnRequest = AuthnRequestBuilder().buildObject().apply {
                 issuer = IssuerBuilder().buildObject().apply {
-                    value = TestCommon.SP_ISSUER
+                    value = SP_ISSUER
                 }
-                id = TestCommon.ID
+                id = ID
                 version = SAMLVersion.VERSION_20
                 issueInstant = DateTime()
-                destination = Common.getSingleSignOnLocation(POST_BINDING)
-                protocolBinding = POST_BINDING
+                destination = Common.getSingleSignOnLocation(SamlProtocol.POST_BINDING)
+                protocolBinding = SamlProtocol.POST_BINDING
                 subject = SubjectBuilder().buildObject()
                 SimpleSign().signSamlObject(this)
             }
@@ -235,15 +152,17 @@ class PostLoginTest : StringSpec() {
             val authnRequestString = authnRequestToString(authnRequest)
             authnRequestString.debugPrettyPrintXml(AUTHN_REQUEST)
 
-            val encodedRequest = Encoder.encodePostMessage(authnRequestString, EXAMPLE_RELAY_STATE)
+            val encodedRequest = Encoder.encodePostMessage(authnRequestString,
+                    EXAMPLE_RELAY_STATE)
             val response = sendAuthnRequest(encodedRequest)
             BindingVerifier.verifyHttpStatusCode(response.statusCode)
 
-            val idpResponse = TestCommon.parseErrorResponse(response)
+            val idpResponse = parseErrorResponse(response)
             idpResponse.bindingVerifier().verifyError()
 
             val responseDom = idpResponse.responseDom
-            CoreVerifier(responseDom).verifyErrorStatusCode(SAMLProfiles_4_1_4_1_b, REQUESTER)
+            CoreVerifier(responseDom).verifyErrorStatusCode(SAMLProfiles_4_1_4_1_b,
+                    REQUESTER)
             ProfilesVerifier(responseDom).verifyErrorResponseAssertion(SAMLProfiles_4_1_4_1_b)
         }.config(enabled = false)
 
@@ -251,13 +170,13 @@ class PostLoginTest : StringSpec() {
             Log.debugWithSupplier { "POST AuthnRequest With Incorrect ACS URL And Index Test" }
             val authnRequest = AuthnRequestBuilder().buildObject().apply {
                 issuer = IssuerBuilder().buildObject().apply {
-                    value = TestCommon.SP_ISSUER
+                    value = SP_ISSUER
                 }
-                id = TestCommon.ID
+                id = ID
                 version = SAMLVersion.VERSION_20
                 issueInstant = DateTime()
-                destination = Common.getSingleSignOnLocation(POST_BINDING)
-                protocolBinding = POST_BINDING
+                destination = Common.getSingleSignOnLocation(SamlProtocol.POST_BINDING)
+                protocolBinding = SamlProtocol.POST_BINDING
                 assertionConsumerServiceURL = INCORRECT_ACS_URL
                 assertionConsumerServiceIndex = -1
                 SimpleSign().signSamlObject(this)
@@ -266,11 +185,12 @@ class PostLoginTest : StringSpec() {
             val authnRequestString = authnRequestToString(authnRequest)
             authnRequestString.debugPrettyPrintXml(AUTHN_REQUEST)
 
-            val encodedRequest = Encoder.encodePostMessage(authnRequestString, EXAMPLE_RELAY_STATE)
+            val encodedRequest = Encoder.encodePostMessage(authnRequestString,
+                    EXAMPLE_RELAY_STATE)
             val response = sendAuthnRequest(encodedRequest)
             BindingVerifier.verifyHttpStatusCode(response.statusCode)
 
-            val idpResponse = TestCommon.parseErrorResponse(response)
+            val idpResponse = parseErrorResponse(response)
             idpResponse.bindingVerifier().verifyError()
 
             val responseDom = idpResponse.responseDom
@@ -281,10 +201,10 @@ class PostLoginTest : StringSpec() {
             Log.debugWithSupplier { "POST AuthnRequest With Non-Matching Destination" }
             val authnRequest = AuthnRequestBuilder().buildObject().apply {
                 issuer = IssuerBuilder().buildObject().apply {
-                    value = TestCommon.SP_ISSUER
+                    value = SP_ISSUER
                 }
                 assertionConsumerServiceURL = acsUrl[SamlProtocol.Binding.HTTP_POST]
-                id = TestCommon.ID
+                id = ID
                 version = SAMLVersion.VERSION_20
                 issueInstant = DateTime()
                 destination = INCORRECT_DESTINATION
@@ -292,7 +212,7 @@ class PostLoginTest : StringSpec() {
                 nameIDPolicy = NameIDPolicyBuilder().buildObject().apply {
                     allowCreate = true
                     format = SAML2Constants.NAMEID_FORMAT_PERSISTENT
-                    spNameQualifier = TestCommon.SP_ISSUER
+                    spNameQualifier = SP_ISSUER
                 }
                 SimpleSign().signSamlObject(this)
             }
@@ -304,11 +224,11 @@ class PostLoginTest : StringSpec() {
             val response = sendAuthnRequest(encodedRequest)
 
             BindingVerifier.verifyHttpStatusCode(response.statusCode)
-            val idpResponse = TestCommon.parseErrorResponse(response)
+            val idpResponse = parseErrorResponse(response)
             idpResponse.bindingVerifier().verifyError()
 
             val responseDom = idpResponse.responseDom
-            CoreVerifier(responseDom).verifyErrorStatusCode(SAMLCore_3_2_1_e, TestCommon.REQUESTER)
+            CoreVerifier(responseDom).verifyErrorStatusCode(SAMLCore_3_2_1_e, REQUESTER)
         }.config(enabled = false)
     }
 }
