+++ ctk/idp/src/main/kotlin/org/codice/compliance/web/sso/error/RedirectSSOErrorTest.kt
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.compliance.web.sso
+package org.codice.compliance.web.sso.error
 
 import com.jayway.restassured.RestAssured
 import com.jayway.restassured.RestAssured.given
@@ -26,21 +26,19 @@ import org.codice.compliance.SAMLProfiles_4_1_4_1_a
 import org.codice.compliance.SAMLProfiles_4_1_4_1_b
 import org.codice.compliance.debugPrettyPrintXml
 import org.codice.compliance.debugWithSupplier
-import org.codice.compliance.saml.plugin.IdpSSOResponder
-import org.codice.compliance.utils.TestCommon
 import org.codice.compliance.utils.TestCommon.Companion.AUTHN_REQUEST
 import org.codice.compliance.utils.TestCommon.Companion.ID
 import org.codice.compliance.utils.TestCommon.Companion.INCORRECT_ACS_URL
 import org.codice.compliance.utils.TestCommon.Companion.INCORRECT_DESTINATION
+import org.codice.compliance.utils.TestCommon.Companion.RELAY_STATE_GREATER_THAN_80_BYTES
+import org.codice.compliance.utils.TestCommon.Companion.REQUESTER
+import org.codice.compliance.utils.TestCommon.Companion.SP_ISSUER
 import org.codice.compliance.utils.TestCommon.Companion.acsUrl
 import org.codice.compliance.utils.TestCommon.Companion.authnRequestToString
-import org.codice.compliance.utils.TestCommon.Companion.getServiceProvider
-import org.codice.compliance.utils.decorate
+import org.codice.compliance.utils.TestCommon.Companion.parseErrorResponse
 import org.codice.compliance.verification.binding.BindingVerifier
 import org.codice.compliance.verification.core.CoreVerifier
-import org.codice.compliance.verification.core.responses.AuthnRequestProtocolResponseVerifier
 import org.codice.compliance.verification.profile.ProfilesVerifier
-import org.codice.compliance.verification.profile.SingleSignOnProfileVerifier
 import org.codice.security.saml.SamlProtocol.Binding.HTTP_POST
 import org.codice.security.saml.SamlProtocol.REDIRECT_BINDING
 import org.codice.security.sign.Encoder
@@ -52,7 +50,7 @@ import org.opensaml.saml.saml2.core.impl.AuthnRequestBuilder
 import org.opensaml.saml.saml2.core.impl.IssuerBuilder
 import org.opensaml.saml.saml2.core.impl.SubjectBuilder
 
-class RedirectLoginTest : StringSpec() {
+class RedirectSSOErrorTest : StringSpec() {
     companion object {
         /**
          * Provides a default request for testing
@@ -61,10 +59,10 @@ class RedirectLoginTest : StringSpec() {
         private fun createDefaultAuthnRequest(): AuthnRequest {
             return AuthnRequestBuilder().buildObject().apply {
                 issuer = IssuerBuilder().buildObject().apply {
-                    value = TestCommon.SP_ISSUER
+                    value = SP_ISSUER
                 }
                 assertionConsumerServiceURL = acsUrl[HTTP_POST]
-                id = TestCommon.ID
+                id = ID
                 version = SAMLVersion.VERSION_20
                 issueInstant = DateTime()
                 destination = Common.getSingleSignOnLocation(REDIRECT_BINDING)
@@ -102,85 +100,6 @@ class RedirectLoginTest : StringSpec() {
     init {
         RestAssured.useRelaxedHTTPSValidation()
 
-        "Redirect AuthnRequest Test" {
-            Log.debugWithSupplier { "Redirect AuthnRequest Test" }
-            val authnRequest = createDefaultAuthnRequest()
-            val encodedRequest = encodeAuthnRequest(authnRequest)
-            val queryParams = SimpleSign().signUriString(
-                    SAML_REQUEST,
-                    encodedRequest,
-                    null)
-
-            val response = sendAuthnRequest(queryParams)
-            BindingVerifier.verifyHttpStatusCode(response.statusCode)
-
-            // Get response from plugin portion
-            val idpResponse = getServiceProvider(IdpSSOResponder::class)
-                    .getRedirectResponse(response).decorate()
-            // TODO When DDF is fixed to return a POST SSO response, uncomment this line
-            // SingleSignOnProfileVerifier.verifyBinding(idpResponse)
-            idpResponse.bindingVerifier().verify()
-
-            val responseDom = idpResponse.responseDom
-            AuthnRequestProtocolResponseVerifier(responseDom, ID, acsUrl[HTTP_POST]).verify()
-            SingleSignOnProfileVerifier(responseDom, acsUrl[HTTP_POST]).verify()
-        }
-
-        "Redirect AuthnRequest With Relay State Test" {
-            Log.debugWithSupplier { "Redirect AuthnRequest With Relay State Test" }
-            val authnRequest = createDefaultAuthnRequest()
-            val encodedRequest = encodeAuthnRequest(authnRequest)
-            val queryParams = SimpleSign().signUriString(
-                    SAML_REQUEST, encodedRequest,
-                    TestCommon.EXAMPLE_RELAY_STATE)
-
-            // Get response from AuthnRequest
-            val response = sendAuthnRequest(queryParams)
-            BindingVerifier.verifyHttpStatusCode(response.statusCode)
-
-            val idpResponse = getServiceProvider(IdpSSOResponder::class)
-                    .getRedirectResponse(response).decorate().apply {
-                        isRelayStateGiven = true
-                    }
-            // TODO When DDF is fixed to return a POST SSO response, uncomment this line
-            // SingleSignOnProfileVerifier.verifyBinding(idpResponse)
-            idpResponse.bindingVerifier().verify()
-
-            val responseDom = idpResponse.responseDom
-            AuthnRequestProtocolResponseVerifier(responseDom, TestCommon.ID, acsUrl[HTTP_POST])
-                    .verify()
-            SingleSignOnProfileVerifier(responseDom, acsUrl[HTTP_POST]).verify()
-        }
-
-        "Redirect AuthnRequest Without ACS Url Test" {
-            Log.debugWithSupplier { "Redirect AuthnRequest Without ACS Url Test" }
-            val authnRequest = createDefaultAuthnRequest().apply {
-                assertionConsumerServiceURL = null
-            }
-
-            val encodedRequest = encodeAuthnRequest(authnRequest)
-            val queryParams = SimpleSign().signUriString(
-                    SAML_REQUEST,
-                    encodedRequest,
-                    null)
-
-            // Get response from AuthnRequest
-            val response = sendAuthnRequest(queryParams)
-            BindingVerifier.verifyHttpStatusCode(response.statusCode)
-
-            // Get response from plugin portion
-            val idpResponse = getServiceProvider(IdpSSOResponder::class)
-                    .getRedirectResponse(response).decorate()
-            // TODO When DDF is fixed to return a POST SSO response, uncomment this line
-            // SingleSignOnProfileVerifier.verifyBinding(idpResponse)
-            idpResponse.bindingVerifier().verify()
-
-            val responseDom = idpResponse.responseDom
-            AuthnRequestProtocolResponseVerifier(responseDom, TestCommon.ID, acsUrl[HTTP_POST])
-                    .verify()
-            SingleSignOnProfileVerifier(responseDom, acsUrl[HTTP_POST]).verify()
-        }
-
         // Negative Path Tests
         "Redirect AuthnRequest With Relay State Greater Than 80 Bytes Test" {
             Log.debugWithSupplier {
@@ -191,18 +110,18 @@ class RedirectLoginTest : StringSpec() {
             val queryParams = SimpleSign().signUriString(
                     SAML_REQUEST,
                     encodedRequest,
-                    TestCommon.RELAY_STATE_GREATER_THAN_80_BYTES)
+                    RELAY_STATE_GREATER_THAN_80_BYTES)
 
             // Get response from AuthnRequest
             val response = sendAuthnRequest(queryParams)
 
-            val idpResponse = TestCommon.parseErrorResponse(response)
+            val idpResponse = parseErrorResponse(response)
             idpResponse.bindingVerifier().verifyError()
 
             val responseDom = idpResponse.responseDom
             CoreVerifier(responseDom).verifyErrorStatusCode(
                     samlErrorCode = SAMLBindings_3_4_3_a1,
-                    expectedStatusCode = TestCommon.REQUESTER)
+                    expectedStatusCode = REQUESTER)
         }.config(enabled = false)
 
         "Redirect Incomplete AuthnRequest In URL Query Test" {
@@ -221,13 +140,13 @@ class RedirectLoginTest : StringSpec() {
             // Get response from AuthnRequest
             val response = sendAuthnRequest(queryParams)
 
-            val idpResponse = TestCommon.parseErrorResponse(response)
+            val idpResponse = parseErrorResponse(response)
             idpResponse.bindingVerifier().verifyError()
 
             val responseDom = idpResponse.responseDom
             CoreVerifier(responseDom).verifyErrorStatusCode(
                     SAMLBindings_3_4_3_a1,
-                    TestCommon.REQUESTER)
+                    REQUESTER)
         }.config(enabled = false)
 
         "Empty Redirect AuthnRequest Test" {
@@ -239,12 +158,12 @@ class RedirectLoginTest : StringSpec() {
             val response = sendAuthnRequest(queryParams)
             BindingVerifier.verifyHttpStatusCode(response.statusCode)
 
-            val idpResponse = TestCommon.parseErrorResponse(response)
+            val idpResponse = parseErrorResponse(response)
             idpResponse.bindingVerifier().verifyError()
 
             val responseDom = idpResponse.responseDom
             CoreVerifier(responseDom).verifyErrorStatusCode(SAMLProfiles_4_1_4_1_a,
-                    TestCommon.REQUESTER)
+                    REQUESTER)
             ProfilesVerifier(responseDom).verifyErrorResponseAssertion()
         }.config(enabled = false)
 
@@ -263,12 +182,12 @@ class RedirectLoginTest : StringSpec() {
             val response = sendAuthnRequest(queryParams)
             BindingVerifier.verifyHttpStatusCode(response.statusCode)
 
-            val idpResponse = TestCommon.parseErrorResponse(response)
+            val idpResponse = parseErrorResponse(response)
             idpResponse.bindingVerifier().verifyError()
 
             val responseDom = idpResponse.responseDom
             CoreVerifier(responseDom).verifyErrorStatusCode(SAMLProfiles_4_1_4_1_b,
-                    TestCommon.REQUESTER)
+                    REQUESTER)
             ProfilesVerifier(responseDom).verifyErrorResponseAssertion(SAMLProfiles_4_1_4_1_b)
         }.config(enabled = false)
 
@@ -285,7 +204,7 @@ class RedirectLoginTest : StringSpec() {
             val response = sendAuthnRequest(queryParams)
             BindingVerifier.verifyHttpStatusCode(response.statusCode)
 
-            val idpResponse = TestCommon.parseErrorResponse(response)
+            val idpResponse = parseErrorResponse(response)
             idpResponse.bindingVerifier().verifyError()
 
             val responseDom = idpResponse.responseDom
@@ -304,11 +223,11 @@ class RedirectLoginTest : StringSpec() {
             val response = sendAuthnRequest(queryParams)
             BindingVerifier.verifyHttpStatusCode(response.statusCode)
 
-            val idpResponse = TestCommon.parseErrorResponse(response)
+            val idpResponse = parseErrorResponse(response)
             idpResponse.bindingVerifier().verifyError()
 
             val responseDom = idpResponse.responseDom
-            CoreVerifier(responseDom).verifyErrorStatusCode(SAMLCore_3_2_1_e, TestCommon.REQUESTER)
+            CoreVerifier(responseDom).verifyErrorStatusCode(SAMLCore_3_2_1_e, REQUESTER)
         }.config(enabled = false)
     }
 }
